respect user config for references and avoid creating a vs code file
