+++ packages/svelte-vscode/src/extension.ts
@@ -29,7 +29,7 @@ import CompiledCodeContentProvider from './CompiledCodeContentProvider';
 import { activateTagClosing } from './html/autoClose';
 import { EMPTY_ELEMENTS } from './html/htmlEmptyTagsShared';
 import { TsPlugin } from './tsplugin';
-import { addFindComponentUsagesListener } from './typescript/findComponentUsages';
+import { addFindComponentReferencesListener } from './typescript/findComponentReferences';
 import { addFindFileReferencesListener } from './typescript/findFileReferences';
 
 namespace TagCloseRequest {
@@ -226,7 +226,7 @@ export function activateSvelteLanguageServer(context: ExtensionContext) {
     addDidChangeTextDocumentListener(getLS);
 
     addFindFileReferencesListener(getLS, context);
-    addFindComponentUsagesListener(getLS, context);
+    addFindComponentReferencesListener(getLS, context);
 
     addRenameFileListener(getLS);
